We are so back
Vargas reminded me about this project, and I decided to restart this project. I rewrote the entire code from scratch and now the current code is actually good, much better than the previous mess,. It's so good that I'll salvage some code from here to Alyssa HTTP Server.

I also reordered the stuff to do. First is completing the base HTTP (which is done already), second is implementing the threading, and third is implementing vectors.