Multithreading implemented
Multithreading is now implemented in a thread pool way.
Connections are polled, and passed to a thread which is waiting for a client.
Unlike AHS or most other mainstream servers which creates a thread per connection and handles from there in blocking way.

Vectors are the next. But not sure when I'll look at it.